Refactor preflight validations into dedicated module
