+++ xtask/src/commands/preflight/validation.rs
@@ -1,0 +1,586 @@
+use crate::error::{TaskError, TaskResult};
+use crate::util::{cargo_metadata_json, ensure, validation_error};
+use crate::workspace::WorkspaceBranding;
+use serde_json::Value as JsonValue;
+use std::collections::HashMap;
+use std::fs;
+use std::path::{Path, PathBuf};
+use toml::{Value, value::Table as TomlTable};
+
+pub(crate) fn validate_branding(branding: &WorkspaceBranding) -> TaskResult<()> {
+    ensure(
+        branding.brand == "oc",
+        format!("workspace brand must be 'oc', found {:?}", branding.brand),
+    )?;
+    ensure(
+        branding.upstream_version == "3.4.1",
+        format!(
+            "upstream_version must remain aligned with rsync 3.4.1; found {:?}",
+            branding.upstream_version
+        ),
+    )?;
+    ensure(
+        branding.rust_version.ends_with("-rust"),
+        format!(
+            "Rust-branded version should end with '-rust'; found {:?}",
+            branding.rust_version
+        ),
+    )?;
+    ensure(
+        branding.protocol == 32,
+        format!("Supported protocol must be 32; found {}", branding.protocol),
+    )?;
+    ensure(
+        branding.client_bin.starts_with("oc-"),
+        format!(
+            "client_bin must start with 'oc-'; found {:?}",
+            branding.client_bin
+        ),
+    )?;
+    ensure(
+        branding.daemon_bin.starts_with("oc-"),
+        format!(
+            "daemon_bin must start with 'oc-'; found {:?}",
+            branding.daemon_bin
+        ),
+    )?;
+
+    let config_dir = Path::new(&branding.daemon_config_dir);
+    ensure(
+        config_dir.is_absolute(),
+        format!(
+            "daemon_config_dir must be an absolute path; found {}",
+            branding.daemon_config_dir
+        ),
+    )?;
+
+    let config_path = Path::new(&branding.daemon_config);
+    let secrets_path = Path::new(&branding.daemon_secrets);
+    ensure(
+        config_path.is_absolute(),
+        format!(
+            "daemon_config must be an absolute path; found {}",
+            branding.daemon_config
+        ),
+    )?;
+    ensure(
+        secrets_path.is_absolute(),
+        format!(
+            "daemon_secrets must be an absolute path; found {}",
+            branding.daemon_secrets
+        ),
+    )?;
+
+    ensure(
+        config_path.parent() == Some(config_dir),
+        format!(
+            "daemon_config {} must reside within configured directory {}",
+            branding.daemon_config, branding.daemon_config_dir
+        ),
+    )?;
+    ensure(
+        secrets_path.parent() == Some(config_dir),
+        format!(
+            "daemon_secrets {} must reside within configured directory {}",
+            branding.daemon_secrets, branding.daemon_config_dir
+        ),
+    )?;
+
+    ensure(
+        config_path.file_name() != secrets_path.file_name(),
+        "daemon configuration and secrets paths must not collide",
+    )?;
+
+    Ok(())
+}
+
+pub(crate) fn validate_packaging_assets(
+    workspace: &Path,
+    branding: &WorkspaceBranding,
+) -> TaskResult<()> {
+    let packaging_root = workspace.join("packaging").join("etc").join("oc-rsyncd");
+    let config_name = file_name(&branding.daemon_config, "daemon_config")?;
+    let secrets_name = file_name(&branding.daemon_secrets, "daemon_secrets")?;
+
+    for (name, label) in [
+        (config_name, "daemon_config"),
+        (secrets_name, "daemon_secrets"),
+    ] {
+        let candidate = packaging_root.join(name);
+        ensure(
+            candidate.exists(),
+            format!(
+                "packaging assets missing for {} (expected {})",
+                label,
+                candidate.display()
+            ),
+        )?;
+    }
+
+    validate_bin_manifest_packaging(workspace, branding)?;
+    validate_systemd_unit(workspace, branding)?;
+    validate_default_env(workspace)?;
+
+    Ok(())
+}
+
+pub(crate) fn validate_package_versions(
+    workspace: &Path,
+    branding: &WorkspaceBranding,
+) -> TaskResult<()> {
+    let metadata = cargo_metadata_json(workspace)?;
+    let packages = metadata
+        .get("packages")
+        .and_then(JsonValue::as_array)
+        .ok_or_else(|| validation_error("cargo metadata output missing packages array"))?;
+
+    let mut versions = HashMap::new();
+    for package in packages {
+        let Some(name) = package.get("name").and_then(JsonValue::as_str) else {
+            continue;
+        };
+        let Some(version) = package.get("version").and_then(JsonValue::as_str) else {
+            continue;
+        };
+        versions.insert(name.to_string(), version.to_string());
+    }
+
+    for crate_name in ["oc-rsync-bin", "oc-rsyncd-bin"] {
+        let version = versions.get(crate_name).ok_or_else(|| {
+            validation_error(format!("crate {crate_name} missing from cargo metadata"))
+        })?;
+        ensure(
+            version == &branding.rust_version,
+            format!(
+                "crate {crate_name} version {version} does not match {}",
+                branding.rust_version
+            ),
+        )?;
+    }
+
+    Ok(())
+}
+
+pub(crate) fn validate_workspace_package_rust_version(manifest: &Value) -> TaskResult<()> {
+    let workspace = manifest
+        .get("workspace")
+        .and_then(Value::as_table)
+        .ok_or_else(|| validation_error("missing [workspace] table in Cargo.toml"))?;
+    let package = workspace
+        .get("package")
+        .and_then(Value::as_table)
+        .ok_or_else(|| validation_error("missing [workspace.package] table in Cargo.toml"))?;
+    let rust_version = package
+        .get("rust-version")
+        .and_then(Value::as_str)
+        .ok_or_else(|| {
+            validation_error("workspace.package.rust-version missing from Cargo.toml")
+        })?;
+
+    ensure(
+        rust_version == "1.87",
+        format!(
+            "workspace.package.rust-version must match CI toolchain 1.87; found {:?}",
+            rust_version
+        ),
+    )
+}
+
+pub(crate) fn validate_documentation(
+    workspace: &Path,
+    branding: &WorkspaceBranding,
+) -> TaskResult<()> {
+    struct DocumentationCheck<'a> {
+        relative_path: &'a str,
+        required_snippets: Vec<&'a str>,
+    }
+
+    let checks = [
+        DocumentationCheck {
+            relative_path: "README.md",
+            required_snippets: vec![
+                branding.client_bin.as_str(),
+                branding.daemon_bin.as_str(),
+                branding.rust_version.as_str(),
+                branding.daemon_config.as_str(),
+            ],
+        },
+        DocumentationCheck {
+            relative_path: "docs/production_scope_p1.md",
+            required_snippets: vec![
+                branding.client_bin.as_str(),
+                branding.daemon_bin.as_str(),
+                branding.rust_version.as_str(),
+                branding.daemon_config_dir.as_str(),
+                branding.daemon_config.as_str(),
+                branding.daemon_secrets.as_str(),
+            ],
+        },
+        DocumentationCheck {
+            relative_path: "docs/differences.md",
+            required_snippets: vec![
+                branding.client_bin.as_str(),
+                branding.daemon_bin.as_str(),
+                branding.rust_version.as_str(),
+            ],
+        },
+        DocumentationCheck {
+            relative_path: "docs/gaps.md",
+            required_snippets: vec![
+                branding.client_bin.as_str(),
+                branding.daemon_bin.as_str(),
+                branding.rust_version.as_str(),
+            ],
+        },
+        DocumentationCheck {
+            relative_path: "docs/COMPARE.md",
+            required_snippets: vec![
+                branding.client_bin.as_str(),
+                branding.daemon_bin.as_str(),
+                branding.rust_version.as_str(),
+                branding.daemon_config.as_str(),
+            ],
+        },
+        DocumentationCheck {
+            relative_path: "docs/feature_matrix.md",
+            required_snippets: vec![
+                branding.client_bin.as_str(),
+                branding.daemon_bin.as_str(),
+                branding.rust_version.as_str(),
+                branding.daemon_config.as_str(),
+                branding.daemon_secrets.as_str(),
+            ],
+        },
+    ];
+
+    for check in checks {
+        let path = workspace.join(check.relative_path);
+        let contents = fs::read_to_string(&path).map_err(|error| {
+            TaskError::Io(std::io::Error::new(
+                error.kind(),
+                format!("failed to read {}: {error}", path.display()),
+            ))
+        })?;
+
+        let missing: Vec<&str> = check
+            .required_snippets
+            .iter()
+            .copied()
+            .filter(|snippet| !snippet.is_empty() && !contents.contains(snippet))
+            .collect();
+
+        ensure(
+            missing.is_empty(),
+            format!(
+                "{} missing required documentation snippets: {}",
+                check.relative_path,
+                missing
+                    .iter()
+                    .map(|snippet| format!("'{}'", snippet))
+                    .collect::<Vec<_>>()
+                    .join(", ")
+            ),
+        )?;
+    }
+
+    Ok(())
+}
+
+pub(crate) fn validate_bin_manifest_packaging(
+    workspace: &Path,
+    branding: &WorkspaceBranding,
+) -> TaskResult<()> {
+    let manifest_path = workspace.join("bin").join("oc-rsync").join("Cargo.toml");
+    let manifest_text = fs::read_to_string(&manifest_path).map_err(|error| {
+        TaskError::Io(std::io::Error::new(
+            error.kind(),
+            format!("failed to read {}: {error}", manifest_path.display()),
+        ))
+    })?;
+
+    let manifest_value: Value = manifest_text.parse().map_err(|error| {
+        TaskError::Metadata(format!(
+            "failed to parse {}: {error}",
+            manifest_path.display()
+        ))
+    })?;
+
+    let package = manifest_value
+        .get("package")
+        .and_then(Value::as_table)
+        .ok_or_else(|| {
+            validation_error(format!(
+                "{} missing [package] table",
+                manifest_path.display()
+            ))
+        })?;
+    let metadata = package
+        .get("metadata")
+        .and_then(Value::as_table)
+        .ok_or_else(|| {
+            validation_error(format!(
+                "{} missing [package.metadata] table",
+                manifest_path.display()
+            ))
+        })?;
+
+    let deb = metadata
+        .get("deb")
+        .and_then(Value::as_table)
+        .ok_or_else(|| {
+            validation_error(format!(
+                "{} missing [package.metadata.deb] table",
+                manifest_path.display()
+            ))
+        })?;
+    ensure(
+        deb_assets_include(deb, branding.daemon_config.as_str()),
+        format!(
+            "{} package.metadata.deb.assets must install {}",
+            manifest_path.display(),
+            branding.daemon_config
+        ),
+    )?;
+    ensure(
+        deb_assets_include(deb, branding.daemon_secrets.as_str()),
+        format!(
+            "{} package.metadata.deb.assets must install {}",
+            manifest_path.display(),
+            branding.daemon_secrets
+        ),
+    )?;
+
+    ensure(
+        deb_conf_files_include(deb, branding.daemon_config.as_str()),
+        format!(
+            "{} package.metadata.deb.conf-files must reference {}",
+            manifest_path.display(),
+            branding.daemon_config
+        ),
+    )?;
+    ensure(
+        deb_conf_files_include(deb, branding.daemon_secrets.as_str()),
+        format!(
+            "{} package.metadata.deb.conf-files must reference {}",
+            manifest_path.display(),
+            branding.daemon_secrets
+        ),
+    )?;
+
+    let rpm = metadata
+        .get("rpm")
+        .and_then(Value::as_table)
+        .ok_or_else(|| {
+            validation_error(format!(
+                "{} missing [package.metadata.rpm] table",
+                manifest_path.display()
+            ))
+        })?;
+    ensure(
+        rpm_assets_include(rpm, branding.daemon_config.as_str()),
+        format!(
+            "{} package.metadata.rpm.assets must install {} with config=true",
+            manifest_path.display(),
+            branding.daemon_config
+        ),
+    )?;
+    ensure(
+        rpm_assets_include(rpm, branding.daemon_secrets.as_str()),
+        format!(
+            "{} package.metadata.rpm.assets must install {} with config=true",
+            manifest_path.display(),
+            branding.daemon_secrets
+        ),
+    )?;
+
+    Ok(())
+}
+
+fn validate_systemd_unit(workspace: &Path, branding: &WorkspaceBranding) -> TaskResult<()> {
+    let systemd_unit = workspace
+        .join("packaging")
+        .join("systemd")
+        .join("oc-rsyncd.service");
+    let unit_contents = fs::read_to_string(&systemd_unit).map_err(|error| {
+        TaskError::Io(std::io::Error::new(
+            error.kind(),
+            format!("failed to read {}: {}", systemd_unit.display(), error),
+        ))
+    })?;
+
+    let unit_snippets = [
+        branding.daemon_bin.as_str(),
+        branding.daemon_config.as_str(),
+        branding.daemon_secrets.as_str(),
+        "Description=oc-rsyncd",
+        "Alias=rsyncd.service",
+        "OC_RSYNC_CONFIG",
+        "OC_RSYNC_SECRETS",
+        "RSYNCD_CONFIG",
+        "RSYNCD_SECRETS",
+    ];
+
+    for snippet in unit_snippets {
+        ensure(
+            unit_contents.contains(snippet),
+            format!(
+                "systemd unit {} missing required snippet '{}': update packaging/systemd/oc-rsyncd.service",
+                systemd_unit.display(),
+                snippet
+            ),
+        )?;
+    }
+
+    Ok(())
+}
+
+fn validate_default_env(workspace: &Path) -> TaskResult<()> {
+    let env_file = workspace
+        .join("packaging")
+        .join("default")
+        .join("oc-rsyncd");
+    let env_contents = fs::read_to_string(&env_file).map_err(|error| {
+        TaskError::Io(std::io::Error::new(
+            error.kind(),
+            format!("failed to read {}: {}", env_file.display(), error),
+        ))
+    })?;
+
+    let env_snippets = [
+        "OC_RSYNC_CONFIG",
+        "RSYNCD_CONFIG",
+        "OC_RSYNC_SECRETS",
+        "RSYNCD_SECRETS",
+    ];
+
+    for snippet in env_snippets {
+        ensure(
+            env_contents.contains(snippet),
+            format!(
+                "environment defaults {} missing '{}': update packaging/default/oc-rsyncd",
+                env_file.display(),
+                snippet
+            ),
+        )?;
+    }
+
+    Ok(())
+}
+
+fn file_name(path: &str, label: &str) -> TaskResult<PathBuf> {
+    let name = Path::new(path)
+        .file_name()
+        .ok_or_else(|| validation_error(format!("{label} must include a file name")))?;
+    Ok(PathBuf::from(name))
+}
+
+fn deb_assets_include(table: &TomlTable, destination: &str) -> bool {
+    table
+        .get("assets")
+        .and_then(Value::as_array)
+        .is_some_and(|assets| {
+            assets.iter().any(|entry| match entry {
+                Value::Array(items) if items.len() >= 2 => {
+                    items.get(1).and_then(Value::as_str) == Some(destination)
+                }
+                _ => false,
+            })
+        })
+}
+
+fn deb_conf_files_include(table: &TomlTable, absolute_path: &str) -> bool {
+    let Some(relative) = absolute_path.strip_prefix('/') else {
+        return false;
+    };
+
+    table
+        .get("conf-files")
+        .and_then(Value::as_array)
+        .is_some_and(|entries| entries.iter().any(|entry| entry.as_str() == Some(relative)))
+}
+
+fn rpm_assets_include(table: &TomlTable, destination: &str) -> bool {
+    table
+        .get("assets")
+        .and_then(Value::as_array)
+        .is_some_and(|assets| {
+            assets.iter().any(|entry| match entry {
+                Value::Table(map) => {
+                    map.get("dest").and_then(Value::as_str) == Some(destination)
+                        && map.get("config").and_then(Value::as_bool).unwrap_or(false)
+                }
+                _ => false,
+            })
+        })
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn deb_asset_helpers_match_expected_entries() {
+        let manifest = r#"
+            [package.metadata.deb]
+            assets = [["source", "/etc/oc-rsyncd/oc-rsyncd.conf", "644"], ["source2", "/etc/oc-rsyncd/oc-rsyncd.secrets", "600"]]
+            conf-files = ["etc/oc-rsyncd/oc-rsyncd.conf", "etc/oc-rsyncd/oc-rsyncd.secrets"]
+        "#;
+        let value: Value = manifest.parse().expect("parse succeeds");
+        let deb = value
+            .get("package")
+            .and_then(Value::as_table)
+            .and_then(|package| package.get("metadata"))
+            .and_then(Value::as_table)
+            .and_then(|metadata| metadata.get("deb"))
+            .and_then(Value::as_table)
+            .expect("deb table present");
+
+        assert!(deb_assets_include(deb, "/etc/oc-rsyncd/oc-rsyncd.conf"));
+        assert!(deb_assets_include(deb, "/etc/oc-rsyncd/oc-rsyncd.secrets"));
+        assert!(deb_conf_files_include(deb, "/etc/oc-rsyncd/oc-rsyncd.conf"));
+        assert!(deb_conf_files_include(
+            deb,
+            "/etc/oc-rsyncd/oc-rsyncd.secrets"
+        ));
+    }
+
+    #[test]
+    fn rpm_asset_helper_requires_config_flag() {
+        let manifest = r#"
+            [package.metadata.rpm]
+            assets = [
+                { path = "src", dest = "/etc/oc-rsyncd/oc-rsyncd.conf", mode = "0644", config = true },
+                { path = "src2", dest = "/etc/oc-rsyncd/oc-rsyncd.secrets", mode = "0600", config = true }
+            ]
+        "#;
+        let value: Value = manifest.parse().expect("parse succeeds");
+        let rpm = value
+            .get("package")
+            .and_then(Value::as_table)
+            .and_then(|package| package.get("metadata"))
+            .and_then(Value::as_table)
+            .and_then(|metadata| metadata.get("rpm"))
+            .and_then(Value::as_table)
+            .expect("rpm table present");
+
+        assert!(rpm_assets_include(rpm, "/etc/oc-rsyncd/oc-rsyncd.conf"));
+        assert!(rpm_assets_include(rpm, "/etc/oc-rsyncd/oc-rsyncd.secrets"));
+
+        let manifest_missing_flag = r#"
+            [package.metadata.rpm]
+            assets = [
+                { path = "src", dest = "/etc/oc-rsyncd/oc-rsyncd.conf", mode = "0644" }
+            ]
+        "#;
+        let value: Value = manifest_missing_flag.parse().expect("parse succeeds");
+        let rpm = value
+            .get("package")
+            .and_then(Value::as_table)
+            .and_then(|package| package.get("metadata"))
+            .and_then(Value::as_table)
+            .and_then(|metadata| metadata.get("rpm"))
+            .and_then(Value::as_table)
+            .expect("rpm table present");
+        assert!(!rpm_assets_include(rpm, "/etc/oc-rsyncd/oc-rsyncd.conf"));
+    }
+}
